Put moved files into the matching namespace
