+++ Samples/Mapsui.Samples.Common/Maps/Tests/BitmapAtlasSample.cs
@@ -1,10 +1,9 @@
 ﻿using Mapsui.Layers;
-using Mapsui.Samples.Common;
 using Mapsui.Styles;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Mapsui.Tests.Common.Maps;
+namespace Mapsui.Samples.Common.Maps.Tests;
 
 public class BitmapAtlasSample : ISample
 {
@@ -12,7 +11,6 @@ public class BitmapAtlasSample : ISample
     public string Category => "Tests";
 
     public Task<Map> CreateMapAsync() => Task.FromResult(CreateMap());
-
 
     public static Map CreateMap()
     {
