+++ Samples/Mapsui.Samples.Common/Maps/Tests/BitmapSymbolSample.cs
@@ -1,18 +1,16 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Mapsui.Layers;
-using Mapsui.Samples.Common;
 using Mapsui.Styles;
 
-namespace Mapsui.Tests.Common.Maps;
+namespace Mapsui.Samples.Common.Maps.Tests;
 
 public class BitmapSymbolSample : ISample
 {
     public string Name => "Bitmap Symbol";
     public string Category => "Tests";
 
-    public Task<Map> CreateMapAsync() => Task.FromResult(BitmapSymbolSample.CreateMap());
-
+    public Task<Map> CreateMapAsync() => Task.FromResult(CreateMap());
 
     public static Map CreateMap()
     {
