+++ Samples/Mapsui.Samples.Common/Maps/Tests/EmptySample.cs
@@ -1,8 +1,7 @@
-﻿using Mapsui.Samples.Common;
-using Mapsui.Styles;
+﻿using Mapsui.Styles;
 using System.Threading.Tasks;
 
-namespace Mapsui.Tests.Common.Maps;
+namespace Mapsui.Samples.Common.Maps.Tests;
 
 public class EmptySample : ISample
 {
@@ -10,7 +9,6 @@ public class EmptySample : ISample
     public string Category => "Tests";
 
     public Task<Map> CreateMapAsync() => Task.FromResult(CreateMap());
-
 
     public static Map CreateMap()
     {
