+++ Samples/Mapsui.Samples.Common/Maps/Tests/LineSample.cs
@@ -2,11 +2,10 @@
 using System.Threading.Tasks;
 using Mapsui.Layers;
 using Mapsui.Nts;
-using Mapsui.Samples.Common;
 using Mapsui.Styles;
 using NetTopologySuite.IO;
 
-namespace Mapsui.Tests.Common.Maps;
+namespace Mapsui.Samples.Common.Maps.Tests;
 
 public class LineSample : ISample
 {
