+++ Samples/Mapsui.Samples.Common/Maps/Tests/PointInWorldUnitsSample.cs
@@ -1,10 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Mapsui.Layers;
-using Mapsui.Samples.Common;
 using Mapsui.Styles;
 
-namespace Mapsui.Tests.Common.Maps;
+namespace Mapsui.Samples.Common.Maps.Tests;
 
 public class PointInWorldUnitsSample : ISample
 {
@@ -12,7 +11,6 @@ public class PointInWorldUnitsSample : ISample
     public string Category => "Tests";
 
     public Task<Map> CreateMapAsync() => Task.FromResult(CreateMap());
-
 
     public static Map CreateMap()
     {
