+++ Samples/Mapsui.Samples.Common/Maps/Tests/SymbolTypesSample.cs
@@ -1,10 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Mapsui.Layers;
-using Mapsui.Samples.Common;
 using Mapsui.Styles;
 
-namespace Mapsui.Tests.Common.Maps;
+namespace Mapsui.Samples.Common.Maps.Tests;
 
 #pragma warning disable IDISP001 // Dispose created
